move ut tests to runtime
